multiple changes, see description
started with common js syntax and provided package.json dependency, but an added module required ES6 syntax so made the necessary changes to it which broke some of the code that was working.

added BUG tag and FIXME tag to, added checklist for what's required
