+++ nodemon.json
@@ -1,4 +1,4 @@
 {
-     "watch": ["server.js", "db", "routes", "public", "public/assets/css", "public/assets/js"],
+     "watch": ["Express.js", "db", "routes", "public", "public/assets/css", "public/assets/js"],
      "ext": "js, html, css, json"
 }
